StockInfo-Backend/requirements.txt
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
pytz==2018.7
xlrd==1.1.0
